+++ scripts/create_sqlite.py
@@ -109,21 +109,8 @@ def add_app_table():
         session.commit()
 
 
-def load_latest():
-    with Session(engine) as session:
-        query = (
-            session.query(AppTable).order_by(AppTable.date_published.desc()).limit(5)
-        )
-        latest = [l[0] for l in session.exec(query).all()]
-
-    with Session(engine) as session:
-        query = session.query(AppTable).order_by(AppTable.date_uploaded.desc()).limit(5)
-        scrape = [l[0] for l in session.exec(query).all()]
-
-    return latest, scrape
-
-
-if __name__ == "__main__":
+def main():
+    #  think this is important!
     papers = add_papers()
 
     for paper in papers.keys():
@@ -133,3 +120,7 @@ if __name__ == "__main__":
             print(f"no JSONLines for {paper}")
 
     add_app_table()
+
+
+if __name__ == "__main__":
+    main()
