+++ templates/layout.html
@@ -30,7 +30,7 @@
 					<p class="muted credit"><a href="/">home</a> -
           <a href="/latest">latest articles</a> -
 					<a href="/random">random article</a> -
-					<a href="/download">download dataset</a> - 
+					<a href="/download">download dataset</a>
         </div>
         {% block body %}{% endblock %}
       </div>
